% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdpalette.R
\name{sdpalette_colour}
\alias{sdpalette_colour}
\title{2 dimensional colourspace}
\usage{
sdpalette_colour(rowvar, colvar)
}
\arguments{
\item{rowvar}{Value for the anomaly in the row variable (in SDs)}

\item{colvar}{Value for the anomaly in the column variable (in SDs)}
}
\value{
a colour as hexidecimal code
}
\description{
Give values from a 2-D colourspace defined by extremeness
(measured in SD) of 2 variables.
}
\details{
This is based on Figure 4 in Babst F, Carrer M, Poulter B,
Urbinati C, Neuwirth B, Frank D. 2012. 500 years of regional
forest growth variability and links to climatic extreme events in
Europe. Environmental Research Letters 7: 45705.
}
