% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdpalette.R
\name{sdpalette_legend}
\alias{sdpalette_legend}
\title{Plot a legend for 2-dimensional colourspace}
\usage{
sdpalette_legend(.xlab = "", .ylab = "")
}
\arguments{
\item{.xlab}{}

\item{.ylab}{}
}
\value{
a gg object
}
